+++ app/src/main/res/menu/menu_detail.xml
@@ -9,4 +9,8 @@
         android:id="@+id/action_delete"
         android:title="@string/eliminar"
         app:showAsAction="never" />
+    <item
+        android:id="@+id/action_logout"
+        android:title="@string/settings_logout"
+        app:showAsAction="never" />
 </menu>
